created util file to hold constant variables that are referenced in various files, moved function get_time_from_file from met_util to string sub, handle legacy env vars for obs_window in separate function so it can be easily removed when we discontinue support for them
