+++ metplus/util/__init__.py
@@ -1,3 +1,4 @@
+from .constants import *
 from .metplus_check import *
 from .doc_util import *
 from .config_metplus import *
