+++ metplus/util/constants.py
@@ -1,0 +1,2 @@
+COMPRESSION_EXTENSIONS = ['.gz', '.bz2', '.zip']
+
